+++ src/cpp11.cpp
@@ -160,6 +160,13 @@ extern "C" SEXP _cheapr_cpp_dbl_sequence(SEXP size, SEXP from, SEXP by) {
   END_CPP11
 }
 // sequences.cpp
+SEXP cpp_sequence(SEXP size, SEXP from, SEXP by);
+extern "C" SEXP _cheapr_cpp_sequence(SEXP size, SEXP from, SEXP by) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_sequence(cpp11::as_cpp<cpp11::decay_t<SEXP>>(size), cpp11::as_cpp<cpp11::decay_t<SEXP>>(from), cpp11::as_cpp<cpp11::decay_t<SEXP>>(by)));
+  END_CPP11
+}
+// sequences.cpp
 SEXP cpp_window_sequence(SEXP size, double k, bool partial, bool ascending);
 extern "C" SEXP _cheapr_cpp_window_sequence(SEXP size, SEXP k, SEXP partial, SEXP ascending) {
   BEGIN_CPP11
@@ -259,6 +266,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_cheapr_cpp_num_na",               (DL_FUNC) &_cheapr_cpp_num_na,               2},
     {"_cheapr_cpp_r_unnested_length",    (DL_FUNC) &_cheapr_cpp_r_unnested_length,    1},
     {"_cheapr_cpp_row_na_counts",        (DL_FUNC) &_cheapr_cpp_row_na_counts,        1},
+    {"_cheapr_cpp_sequence",             (DL_FUNC) &_cheapr_cpp_sequence,             3},
     {"_cheapr_cpp_vec_length",           (DL_FUNC) &_cheapr_cpp_vec_length,           1},
     {"_cheapr_cpp_which_",               (DL_FUNC) &_cheapr_cpp_which_,               2},
     {"_cheapr_cpp_which_na",             (DL_FUNC) &_cheapr_cpp_which_na,             1},